feat(main.py, test.excalidraw and .gitignore): add early attempt to build a weather cli python program
this commit adds a very early stage of a python program that is
intended to be a weather cli program. it is not yet functional and is
not yet tested. the programis intended to be run from the command line
and will take a city name as an argument. the program is intended to
return the current weather for that city